I think I got it totally working but still missing things he wants us to implement - BW
So, if we start with the puzzle from the example and run it, we get the correct result in the correct amount of moves. But we still need to implement the hash table for the explored set AND the FIFO tie breaker.

In the example, idk what to do about when it branches and appears to be executing the two brances at once and then chooses the left one once it becomes smaller... if that makes sense?